solution/openai_agents_mvp/requirements.txt
openai-agents==0.9.2
openai==2.21.0
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.1
aiosqlite==0.22.1
